requirements.txt
flask==2.0.1
flask-cors==3.0.10
faiss-cpu==1.7.2
numpy==1.22.0
requests==2.27.1
beautifulsoup4==4.10.0
markdownify==0.11.6
filelock==3.8.0
python-dotenv==0.20.0
# For Google Gemini (optional)
google-generativeai==0.3.0 